Add new toggle and new flex-based wrappers.
Ad toggle:
In mobile state, only the icons are visible in the sidebar navigation. This toggle – positioned fixed in the bottom left over the sidebar menu and only visible in mobile state – enables the user to see the nav-items' texts that are hidden in mobile state.

Ad flex-based wrappers:
Enable equal-height of sidebar and content. (Sidebar always has its blue background.)
Currently, though, this messes with the styling of the main content a little bit (e.g. footer). This will be handled in a separate PR dedicated to the mobile view of he main content area.