CSS for main navbar and its new search toggle
Move the already existing CSS rule for .main-navbar (`.aimeos .main-navbar span.placeholder`) to the very top, since the .main-navbar is also the first element on page. (There are no preceding rules for placeholder that would affect the result of this change.)

Add new rules for the new search toggle (only visible in mobile state).