For Discussion: Improved mobile layout for sidebar (#125)
* "Toggle nav items texts" function

Adds a class to the (iframe's) body tag, to which CSS can react.

Question: since iframe is AFAIK only a "thing" in TYPO3, is this solution (adding a class to the body tag) still viable for other integrations like Laravel etc.?

* Add new toggle and new flex-based wrappers.

Ad toggle:
In mobile state, only the icons are visible in the sidebar navigation. This toggle – positioned fixed in the bottom left over the sidebar menu and only visible in mobile state – enables the user to see the nav-items' texts that are hidden in mobile state.

Ad flex-based wrappers:
Enable equal-height of sidebar and content. (Sidebar always has its blue background.)
Currently, though, this messes with the styling of the main content a little bit (e.g. footer). This will be handled in a separate PR dedicated to the mobile view of he main content area.

* Disable horizontal scrolling

Currently the navigation moves out of sight, when the main content is scrolled horizontally. IMO it is better to assign horizontal scrolling only to the main content (table) only. That way also the header remains in place and the user always knows where she is.

* CSS for toggle and sidemenu with visible texts

This CSS makes the names/titles of the nav items visible. In order to have enough space, some things are changed in size (e.g. the logo). Otherwise the sub-resource tree-menu won't have enough space on very small devices.

The sub-resource tree-menus adapt to the "nav texts visible" state. Screenshots will be added.

* Fix padding bottom of sidebar-menu.advanced

* Align nav items texts; add vendor prefixes

When "show nav items texts" is active, the icons are assigned a width of 2.25rem, so that all the texts align nicely.

All comments (except license) removed.

The whole file was sent through autoprefixer with these settings:
/*
* Prefixed by https://autoprefixer.github.io
* PostCSS: v7.0.29,
* Autoprefixer: v9.7.6
* Browsers: last 4 version
*/

* Fix code formatting; remove comments

* Change chevron icon to already in-use less icon

The "show nav items texts" toggle introduced a new icon, but since a similar icon is already in use for the "more/less" separator, let's reuse that one.

Also, apply the slighty darker background to the separator and the toggle to distinguish more clearly between the functions of these icons compared to the others.
The toggle expands, when active, and therefore has more similarity with the separator.