Improved mobile layout  main content area (#126)
* CSS for main navbar and its new search toggle

Move the already existing CSS rule for .main-navbar (`.aimeos .main-navbar span.placeholder`) to the very top, since the .main-navbar is also the first element on page. (There are no preceding rules for placeholder that would affect the result of this change.)

Add new rules for the new search toggle (only visible in mobile state).

* Add new toggle to show/hide search

The new toggle is only visible in mobile view and shows/hides the search input fields, which are hidden in mobile view.

* Add function to toggle search

This function adds (and removes) the class `.js--show-search` to the (iframe's) body tag. Dedicated CSS shows/hides the search.

* Add a responsive "current page" display variant.

When in mobile state, only '1/2' is shown.

* Fix pagination CSS. Remove main-navbar styles.

Switch to "mobile first" principle.
Move main-navbar styles to nav.css.

* Add styles for search toggle and main-navbar.

Move footer styles to bottom of file.

Question: Why is this element called "main-navbar"? There is no navigation, "just" a header, a search field and a pagination... It seems I am missing something, but I also cannot find a navbar-secondary anywhere on the Aimeos admin interface.

* More CSS for the search box.

in mobile state, the search box is hidden. Only when the a specific class is added to the body via JS, then some CSS reacts and the search box slides in from the top, resting over the pagination, which, most likely, at this point the user won't need.

* Increase top value for active search box

Using this value, the search box is properly centered between the DOM elements above and below it, when active.

* Reduce width for search icon toggle.

* Remove chevron from search toggle icon.

* Remove merging conflict

The "toggleNavItemsTexts" function does not belong to this PR. Removed.

* Reduce with of "Number per page" selector items 

The dropdown items of the "number per page" selector is reduced and the text centered in order to fit better into desktop and mobile state.
Requires ` dropdown-menu-right` to be available on `<ul class="dropdown-menu">` in `pagination-standard.php`.

* Add "dropdown-menu-right"

This class improves the layout of the "items per page" selector dropdown in mobile state, where the container's dimensions broke out to the right side of the page.

* Fix searchbox positioning and animation

Use `position: absolute` instead of `position: fixed` in order to get rid of the problem regarding scrollbar width on desktop vs mobile browsers / devices.

Use `translateY()` for animation instead of `top`.

Remove unused CSS that was previously introduced, when the search toggle additionally had a chevron icon.