Remove pointer-events, fix box-shadow, ...
... more:
add `padding-bottom: 240px` to `sidebar-menu.advanced` to cover the potential visibility of the sidebar's animation on screens that are higher than all the sidebar menues together. The amount of 240px should be sensible enough to handle 1920x1440 monitors with Aimeos in desktop state and standard setup (not in mobile state, though); at the same time it does not cause a too long scrollbar (since the additional height adds to that, too).