[For Discussion] Position fixed with extra scrollbars (#128)
* Add new div wrappers

Enables sticky headers in sub menues and better control over scrollbars.

* New CSS for "position: fixed" solutions

- Logo sticky
- Headers in submenues sticky
- submenu is fixed
- submenus has its own scrollbar
- native scrollbar styling for webkit Browsers and Firefox 

The structure of the CSS was changed in order to group elements better.

* Add pointer-events rules

The `tree-menu`s must be set to `pointer-events: none`, that disables interactivity with the mouse and therefore removes the flickering on mouse-out.
During the `:hover` state, `pointer-events: all` is set.

* Remove commented-out code

* Remove pointer-events, fix box-shadow, ...

... more:
add `padding-bottom: 240px` to `sidebar-menu.advanced` to cover the potential visibility of the sidebar's animation on screens that are higher than all the sidebar menues together. The amount of 240px should be sensible enough to handle 1920x1440 monitors with Aimeos in desktop state and standard setup (not in mobile state, though); at the same time it does not cause a too long scrollbar (since the additional height adds to that, too).