Update default light / dark theme handling
Updated approach for setting the default light / dark theme to explicitly remove the "light", and add the "dark", class to the body element. Theme is set to "light" by default based on cookie presence / contents.

Note: Previous approach would revert to light theme when browser was set to dark in the applications preferences.